+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Distinguish `ResultsSet` from `ResultsList`.
 - Remove `values_type` parameter
 - `ResultsList` supports `[]`
+- Expose stub pydoc.
 
 0.0.5 - 2021-02-06
 - Support queries for and by dataset